backend/requirements.txt
# =============================================
# FATURA OTOMASYON SİSTEMİ - Backend Dependencies
# =============================================

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0

# Pydantic (Data validation)
pydantic[email]==2.6.1
pydantic-settings==2.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# AI - Claude API
anthropic==0.20.0

# Email - Gmail IMAP + SMTP Notifications
# (imaplib built-in, no install needed)
sendgrid==6.11.0

# File Parsing
PyPDF2==3.0.1
lxml==5.1.0
pymupdf>=1.24.0

# Scheduler (Email polling + Reminder)
apscheduler==3.10.4

# HTTP Client
httpx==0.27.0

# Logging
loguru==0.7.2

# Environment Variables
python-dotenv==1.0.1

# Date/Time
python-dateutil==2.9.0

# JSON serialization
orjson==3.9.15

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
